requirements.txt
requests~=2.25.1
uuid~=1.30
python-decouple~=3.4